git-svn-id: http://qon.jwdt.org/var/svn/catracker@183 97746f1a-7bc6-4ed5-8832-28bd89886fd8

+++ GPU/TPCCAGPUTracking/tracking-ca/standalone/code/AliHLTTPCCASliceData.cxx
@@ -96,8 +96,8 @@ inline void AliHLTTPCCASliceData::PackHitData( AliHLTTPCCARow* const row, const 
       std::cout << "!!!! hit packing error!!! " << xx << " " << yy << " " << std::endl;
     }
     // HitData is bin sorted
-    fHitData[row->fHitNumberOffset + hitIndex].x = xx;
-    fHitData[row->fHitNumberOffset + hitIndex].y = yy;
+    fHitData[row->fHitNumberOffset + hitIndex].x = (unsigned short) xx;
+    fHitData[row->fHitNumberOffset + hitIndex].y = (unsigned short) yy;
   }
 }
